+++ tower-http-client/src/client/service_ext.rs
@@ -3,7 +3,7 @@ use std::future::Future;
 use http::{Method, Uri};
 use tower_service::Service;
 
-use crate::{request_builder::ClientRequest, IntoUri};
+use super::{ClientRequest, IntoUri};
 
 /// An extension trait for Tower HTTP services with the typical client methods.
 ///
@@ -34,8 +34,8 @@ pub trait ServiceExt<ReqBody, RespBody, Err>: Sized {
     where
         ReqBody: Default,
         U: IntoUri,
-        Uri: TryFrom<U::Input>,
-        <Uri as TryFrom<U::Input>>::Error: Into<http::Error>,
+        Uri: TryFrom<U::TryInto>,
+        <Uri as TryFrom<U::TryInto>>::Error: Into<http::Error>,
     {
         ClientRequest::builder(self).method(method).uri(uri)
     }
@@ -45,8 +45,8 @@ pub trait ServiceExt<ReqBody, RespBody, Err>: Sized {
     where
         ReqBody: Default,
         U: IntoUri,
-        Uri: TryFrom<U::Input>,
-        <Uri as TryFrom<U::Input>>::Error: Into<http::Error>,
+        Uri: TryFrom<U::TryInto>,
+        <Uri as TryFrom<U::TryInto>>::Error: Into<http::Error>,
     {
         self.request(Method::GET, uri)
     }
@@ -56,8 +56,8 @@ pub trait ServiceExt<ReqBody, RespBody, Err>: Sized {
     where
         ReqBody: Default,
         U: IntoUri,
-        Uri: TryFrom<U::Input>,
-        <Uri as TryFrom<U::Input>>::Error: Into<http::Error>,
+        Uri: TryFrom<U::TryInto>,
+        <Uri as TryFrom<U::TryInto>>::Error: Into<http::Error>,
     {
         self.request(Method::PUT, uri)
     }
@@ -67,8 +67,8 @@ pub trait ServiceExt<ReqBody, RespBody, Err>: Sized {
     where
         ReqBody: Default,
         U: IntoUri,
-        Uri: TryFrom<U::Input>,
-        <Uri as TryFrom<U::Input>>::Error: Into<http::Error>,
+        Uri: TryFrom<U::TryInto>,
+        <Uri as TryFrom<U::TryInto>>::Error: Into<http::Error>,
     {
         self.request(Method::POST, uri)
     }
@@ -78,8 +78,8 @@ pub trait ServiceExt<ReqBody, RespBody, Err>: Sized {
     where
         ReqBody: Default,
         U: IntoUri,
-        Uri: TryFrom<U::Input>,
-        <Uri as TryFrom<U::Input>>::Error: Into<http::Error>,
+        Uri: TryFrom<U::TryInto>,
+        <Uri as TryFrom<U::TryInto>>::Error: Into<http::Error>,
     {
         self.request(Method::PATCH, uri)
     }
@@ -89,8 +89,8 @@ pub trait ServiceExt<ReqBody, RespBody, Err>: Sized {
     where
         ReqBody: Default,
         U: IntoUri,
-        Uri: TryFrom<U::Input>,
-        <Uri as TryFrom<U::Input>>::Error: Into<http::Error>,
+        Uri: TryFrom<U::TryInto>,
+        <Uri as TryFrom<U::TryInto>>::Error: Into<http::Error>,
     {
         self.request(Method::DELETE, uri)
     }
@@ -100,8 +100,8 @@ pub trait ServiceExt<ReqBody, RespBody, Err>: Sized {
     where
         ReqBody: Default,
         U: IntoUri,
-        Uri: TryFrom<U::Input>,
-        <Uri as TryFrom<U::Input>>::Error: Into<http::Error>,
+        Uri: TryFrom<U::TryInto>,
+        <Uri as TryFrom<U::TryInto>>::Error: Into<http::Error>,
     {
         self.request(Method::HEAD, uri)
     }